Created CameraController Script for camera jolt effect - Mark
Created CameraController Script for camera jolt effect, and set the
Ball_Controller Script to trigger on the event of a goal. This is to add
player feedback when scoring and make the goals feel more powerful. All
varibles can be tweaked in Unity Editor to tweak the effect (in case it
needs to feel more or less subtle). - Mark