Fixxed project file structure and made a build - Mark
We had all the Unity Project file structure a little messed up as
everything was one folder too deep inside a duplicate 'Assets' folder.
So I've moved everything to where it should be and that should fix some
of the little problems and bugs we were expeirencing the other day.

Everything seems to be working fine but would suggest we all download a
fresh copy of the Github repo as changing file structures don't also
update on the local copy (may be fine though) @CCrewe @s160210
@ThatGingerJust